LPS-76089 Moves access policy to model, like class that extends
